Extracts UserPresence into class
# Conflicts:
#	app/server/lib/DocClients.ts
