+++ app/server/lib/DocClients.ts
@@ -3,17 +3,14 @@
  * open, and what FD they are using.
  */
 
-import {VisibleUserProfile} from 'app/common/ActiveDocAPI';
-import {CommDocEventType, CommDocUserPresenceUpdate, CommMessage} from 'app/common/CommTypes';
+import {CommDocEventType, CommMessage} from 'app/common/CommTypes';
 import {arrayRemove} from 'app/common/gutil';
-import * as roles from 'app/common/roles';
-import {getRealAccess} from 'app/common/UserAPI';
 import {ActiveDoc} from 'app/server/lib/ActiveDoc';
 import {appSettings} from 'app/server/lib/AppSettings';
 import {Client} from 'app/server/lib/Client';
 import {DocSession, DocSessionPrecursor} from 'app/server/lib/DocSession';
 import {LogMethods} from "app/server/lib/LogMethods";
-import {fromPairs} from 'lodash';
+import EventEmitter from 'events';
 
 const isUserPresenceEnabledByDefault = false;
 export const SETTING_ENABLE_USER_PRESENCE = appSettings.section('userPresence').flag('enable');
@@ -31,19 +28,32 @@ export const Deps = {
   BROADCAST_ORDER: 'parallel' as 'parallel' | 'series',
 };
 
-export class DocClients {
+export type ClientAddedEventListener = (session: DocSession) => void;
+export type ClientRemovedEventListener = (session: DocSession) => void;
+
+export class DocClients extends EventEmitter {
   private _docSessions: DocSession[] = [];
   private _log = new LogMethods('DocClients ', (s: DocSession|null) => this.activeDoc.getLogMeta(s));
 
   constructor(
     public readonly activeDoc: ActiveDoc
-  ) {}
+  ) {
+    super();
+  }
 
   /**
    * Returns the number of connected clients.
    */
   public clientCount(): number {
     return this._docSessions.length;
+  }
+
+  /**
+   * Lists all active doc sessions
+   * @returns {DocSession[]}
+   */
+  public listClients(): DocSession[] {
+    return this._docSessions.slice();
   }
 
   /**
@@ -54,8 +64,12 @@ export class DocClients {
     this._docSessions.push(docSession);
     this._log.debug(docSession, "now %d clients; new client is %s (fd %s)",
       this._docSessions.length, client.clientId, docSession.fd);
-    this._broadcastUserPresenceSessionUpdate(docSession);
+    this._emitClientAdded(docSession);
     return docSession;
+  }
+
+  public addClientAddedListener(listener: ClientAddedEventListener) {
+    this.on('clientAdded', listener);
   }
 
   /**
@@ -70,7 +84,11 @@ export class DocClients {
       this._log.debug(docSession, "now %d clients", this._docSessions.length);
     }
 
-    this._broadcastUserPresenceSessionRemoval(docSession);
+    this._emitClientRemoved(docSession);
+  }
+
+  public addClientRemovedListener(listener: ClientRemovedEventListener) {
+    this.on('clientRemoved', listener);
   }
 
   /**
@@ -81,7 +99,7 @@ export class DocClients {
     const docSessions = this._docSessions.splice(0);
     for (const docSession of docSessions) {
       docSession.client.removeDocSession(docSession.fd);
-      this._broadcastUserPresenceSessionRemoval(docSession);
+      this._emitClientRemoved(docSession);
     }
   }
 
@@ -90,16 +108,6 @@ export class DocClients {
     for (const docSession of this._docSessions) {
       docSession.client.interruptConnection();
     }
-  }
-
-  public async listVisibleUserProfiles(viewingDocSession: DocSession): Promise<VisibleUserProfile[]> {
-    if (isUserPresenceDisabled()) { return []; }
-    const otherDocSessions = this._docSessions.filter(s => s.client.clientId !== viewingDocSession.client.clientId);
-    const docUserRoles = await this._getDocUserRoles();
-    const userProfiles = otherDocSessions.map(
-      s => getVisibleUserProfileFromDocSession(s, viewingDocSession, docUserRoles)
-    );
-    return userProfiles.filter((s?: VisibleUserProfile): s is VisibleUserProfile => s !== undefined);
   }
 
   /**
@@ -177,97 +185,12 @@ export class DocClients {
     }
   }
 
-  private _broadcastUserPresenceSessionUpdate(originSession: DocSession) {
-    if (isUserPresenceDisabled()) { return; }
-    // Loading the doc user roles first allows the callback to be quick + synchronous,
-    // avoiding a potentially linear series of async calls.
-    this._getDocUserRoles()
-      .then(docUserRoles => this.broadcastDocMessage(
-        originSession.client,
-        "docUserPresenceUpdate",
-        undefined,
-        async (destSession: DocSession, messageData: any): Promise<CommDocUserPresenceUpdate["data"] | undefined> => {
-          if (originSession === destSession) { return; }
-          const profile = getVisibleUserProfileFromDocSession(originSession, destSession, docUserRoles);
-          if (!profile) { return; }
-          return {
-            id: getVisibleUserProfileId(originSession),
-            profile
-          };
-        }
-      ))
-      .catch(err => {
-        this._log.error(originSession, "failed to broadcast user presence session update: %s", err);
-      });
+  private _emitClientAdded(session: DocSession) {
+    this.emit('clientAdded', session);
   }
 
-  private _broadcastUserPresenceSessionRemoval(originSession: DocSession) {
-    if (isUserPresenceDisabled()) { return; }
-    this.broadcastDocMessage(
-      originSession.client,
-      "docUserPresenceUpdate",
-      undefined,
-      async (destSession: DocSession, messageData: any): Promise<CommDocUserPresenceUpdate["data"] | undefined> => {
-        return {
-          id: getVisibleUserProfileId(originSession),
-          profile: null,
-        };
-      }
-    ).catch(err => {
-      this._log.error(originSession, "failed to broadcast user presence session removal: %s", err);
-    });
-  }
-
-  private async _getDocUserRoles(): Promise<UserIdRoleMap> {
-    const homeDb = this.activeDoc.getHomeDbManager();
-    const authCache = homeDb?.caches;
-    const docId = this.activeDoc.doc?.id;
-
-    // Not enough information - no useful data to be had here.
-    if (!homeDb || !docId || !authCache) {
-      return {};
-    }
-
-    const queryResult = await authCache.getDocAccess(docId);
-    const { users, maxInheritedRole } = homeDb.unwrapQueryResult(queryResult);
-
-    return fromPairs(users.map(user => [user.id, getRealAccess(user, { maxInheritedRole })]));
+  private _emitClientRemoved(session: DocSession) {
+    this.emit('clientRemoved', session);
   }
 }
 
-interface UserIdRoleMap {
-  [id: string]: roles.Role | null
-}
-
-function getVisibleUserProfileFromDocSession(
-  session: DocSession, viewingSession: DocSession, docUserRoles: UserIdRoleMap,
-): VisibleUserProfile | undefined {
-  // To see other users, you need to be a non-public user (i.e. added to the document), and have
-  // at least editor permissions.
-  if (!viewingSession.client.authSession.userId) {
-    return undefined;
-  }
-
-  const viewerRole = docUserRoles[viewingSession.client.authSession.userId];
-  if (!roles.canEdit(viewerRole)) {
-    return undefined;
-  }
-
-  const user = session.client.authSession.fullUser;
-  const userId = session.client.authSession.userId;
-  const explicitUserRole = userId ? docUserRoles[userId] : null;
-  // Only signed-in users that have explicit document access or are a member of the org / workspace
-  // have visible details by default.
-  const isAnonymous = !explicitUserRole;
-  return {
-    id: getVisibleUserProfileId(session),
-    name: (isAnonymous ? "Anonymous User" : user?.name) || "Unknown User",
-    email: isAnonymous ? undefined : user?.email,
-    picture: isAnonymous ? undefined : user?.picture,
-    isAnonymous,
-  };
-}
-
-function getVisibleUserProfileId(session: DocSession): string {
-  return session.client.publicClientId;
-}
